Heretic Compiles and changed C to CPP headers

+++ src/heretic/r_segs.cpp
@@ -22,7 +22,7 @@
 //**
 //**************************************************************************
 
-#include <stdlib.h>
+#include <cstdlib>
 
 #include "doomdef.hpp"
 #include "i_system.hpp" // [crispy] I_Realloc()
@@ -372,7 +372,7 @@ void R_StoreWallRange(int start, int stop)
 // calculate rw_distance for scale calculation
 //
     rw_normalangle = curline->angle + ANG90;
-    offsetangle = abs(rw_normalangle - rw_angle1);
+    offsetangle = abs(static_cast<int>(rw_normalangle) - rw_angle1);
     if (offsetangle > ANG90)
         offsetangle = ANG90;
     distangle = ANG90 - offsetangle;
